Cleint side routes for user management
