+++ package.json
@@ -6,7 +6,7 @@
   "dependencies": {
     "firebase": "^4.8.0",
     "gatsby": "^1.9.127",
-    "gatsby-link": "^1.6.30",
+    "gatsby-link": "^1.6.41",
     "gatsby-plugin-react-helmet": "^1.0.8",
     "gatsby-plugin-react-next": "^1.0.11",
     "prop-types": "^15.6.0"
