+++ src/components/Home/index.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 
-import withAuthorization from '../components/Session/withAuthorization';
-import { db } from '../firebase';
+import withAuthorization from '../Session/withAuthorization';
+import { db } from '../../firebase';
 
 const fromObjectToList = (object) =>
   object
