+++ src/constants/routes.js
@@ -1,6 +1,6 @@
 export const LANDING = '/';
-export const SIGN_UP = '/signup';
-export const SIGN_IN = '/signin';
+export const SIGN_UP = '/app/signup';
+export const SIGN_IN = '/app/signin';
 export const PASSWORD_FORGET = '/app/pw-forget';
 export const HOME = '/app/home';
 export const ACCOUNT = '/app/account';
