+++ src/pages/app/index.js
@@ -1,30 +1,25 @@
 import React from 'react';
 import {
-  BrowserRouter as Router,
+  // BrowserRouter as Router,
   Route,
 } from 'react-router-dom';
 
-import Navigation from '../Navigation';
-import LandingPage from '../Landing';
-import SignUpPage from '../SignUp';
-import SignInPage from '../SignIn';
-import PasswordForgetPage from '../PasswordForget';
-import HomePage from '../Home';
-import AccountPage from '../Account';
-import withAuthentication from '../Session/withAuthentication';
+import SignUpPage from '../../components/SignUp';
+import SignInPage from '../../components/SignIn';
+import PasswordForgetPage from '../../components/PasswordForget';
+import HomePage from '../../components/Home';
+import AccountPage from '../../components/Account';
+import withAuthentication from '../../components/Session/withAuthentication';
 import * as routes from '../../constants/routes';
 
 import './index.css';
 
-const App = () =>
-  <Router>
+const app = () =>
+  <div>
     <div className="app">
-      <Navigation />
 
       <hr/>
 
-      <Route exact path={routes.LANDING} component={() => <LandingPage />} />
-      <Route exact path={routes.SIGN_UP} component={() => <SignUpPage />} />
       <Route exact path={routes.SIGN_IN} component={() => <SignInPage />} />
       <Route exact path={routes.PASSWORD_FORGET} component={() => <PasswordForgetPage />} />
       <Route exact path={routes.HOME} component={() => <HomePage />} />
@@ -34,6 +29,6 @@ const App = () =>
 
       <span>Found in <a href="https://roadtoreact.com/course-details?courseId=TAMING_THE_STATE">Taming the State in React</a></span> | <span>Star the <a href="https://github.com/rwieruch/react-firebase-authentication">Repository</a></span> | <span>Receive a <a href="https://www.getrevue.co/profile/rwieruch">Developer's Newsletter</a></span>
     </div>
-  </Router>
+  </div>
 
-export default withAuthentication(App);
+export default withAuthentication(app);
