--- src/pages/pw-forget.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import { PasswordForgetForm } from '../components/PasswordForget';
-
-const PasswordForgetPage = () =>
-  <div>
-    <h1>PasswordForget</h1>
-    <PasswordForgetForm />
-  </div>
-
-export default PasswordForgetPage;
