--- src/pages/signin.js
@@ -1,15 +1,0 @@
-import React from 'react';
-
-import SignInForm from '../components/SignIn';
-import { SignUpLink } from '../components/SignUp';
-import { PasswordForgetLink } from '../components/PasswordForget';
-
-const SignInPage = () =>
-  <div>
-    <h1>SignIn</h1>
-    <SignInForm />
-    <PasswordForgetLink />
-    <SignUpLink />
-  </div>
-
-export default SignInPage;
