--- src/pages/signup.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import SignUpForm from '../components/SignUp';
-
-const SignUpPage = () =>
-  <div>
-    <h1>SignUp</h1>
-    <SignUpForm />
-  </div>
-
-export default SignUpPage;
